Extract for loop generation related code
